Audit Phase D - Multiple Crash protections/fixes.

+++ Server/src/udb_achunk.c
@@ -301,7 +301,7 @@ dddb_mark(off_t lbn, int siz, int taken)
 		last_free = lbn;
 
 	while(bcnt--) {
-		if(lbn >= bm_top - 32)
+		if((int)lbn >= (int)bm_top - 32)
 			growbit(lbn + DDDB_BITBLOCK);
 
 		if(taken)
@@ -330,7 +330,7 @@ dddb_alloc(int siz)
 	bcnt = BLOCKS_NEEDED(siz);
 
 	while(1) {
-		if(lbn >= bm_top - 32) {
+		if((int)lbn >= (int)bm_top - 32) {
 			/* only check here. can't break around the top */
 			if(!overthetop) {
 				lbn = 0;
@@ -357,7 +357,7 @@ dddb_alloc(int siz)
 			}
 
 			lbn++;
-			if(lbn >= bm_top - 32)
+			if((int)lbn >= (int)bm_top - 32)
 				growbit(lbn + DDDB_BITBLOCK);
 		}
 		lbn++;
